backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
gunicorn==21.2.0
requests==2.31.0
pymongo==4.4.0
python-dotenv==1.0.0